EAE 1410: Player Class
This is a class I created as part of a final project for my Intro to Object Oriented Programming Class. The class is that of the playable object which can change shape when certain pickups are collected, pass through holes in walls if the player is the right shape, and move up and down a set of fixed lanes in the game.